feat(sales): replace order creation dialog with dedicated pages
